Corrected logic errors.
Corrected logic errors where 'not in alma' and 'unparsable call
number' were not being shown when limiting the report to only
other issues.

Additionally, corrected logic issue where scanning in was still
active for reports that only were looking for order issues. This
would cause scan in results to not display on the report and
just shouldn't really happen in this mode.