Implement Longest Increasing Subsequence algorithm.
The algorithm that showed the order issues to the user (not the
sorting algorithm, but the one to display the message) was using
a wonky home-grown method that I didn't really understand, even
though I made the whole thing from scratch. There were some weird
behaviors.

Gemini suggested that a Longest Increasing Subsequence algorithm
would be a good fit for this problem. After studying this a little,
I agreed. This code implements the new algorithm, which works
a lot more consistently.

https://umd-dit.atlassian.net/browse/LIBILS-1027